Move the imageutils files to the image package
