+++ cmd/lifecycle/rebaser.go
@@ -17,7 +17,6 @@ import (
 	"github.com/buildpacks/lifecycle/cmd"
 	"github.com/buildpacks/lifecycle/encoding"
 	"github.com/buildpacks/lifecycle/image"
-	"github.com/buildpacks/lifecycle/imageutils"
 	"github.com/buildpacks/lifecycle/platform"
 	"github.com/buildpacks/lifecycle/platform/common"
 	"github.com/buildpacks/lifecycle/priv"
@@ -169,7 +168,7 @@ func (r *rebaseCmd) setAppImage() error {
 	}
 
 	var md platform.LayersMetadata
-	if err := imageutils.DecodeLabel(r.appImage, platform.LayerMetadataLabel, &md); err != nil {
+	if err := image.DecodeLabel(r.appImage, platform.LayerMetadataLabel, &md); err != nil {
 		return err
 	}
